requirements.txt
pandas
sqlalchemy
pymysql
python-dotenv
pwinput